lesson-07-rag/demo/shipping/requirements.txt
a2a-sdk>=0.3.6
google-adk>=1.17.0
toolbox-core>=0.5.0
google-cloud-discoveryengine>=0.13.2